refactor: activity package 구조 변경
